backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
asn1crypto==1.5.1
attrs==25.3.0
boto3==1.39.4
botocore==1.39.4
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.116.1
filelock==3.18.0
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
iniconfig==2.1.0
jiter==0.10.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-openai==0.3.27
langchain-text-splitters==0.3.8
langsmith==0.4.5
marshmallow==3.26.1
multidict==6.6.3
mypy_extensions==1.1.0
numpy==2.3.1
openai==1.95.1
orjson==3.10.18
packaging==24.2
pandas==2.3.1
platformdirs==4.3.8
pluggy==1.6.0
propcache==0.3.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.1.0
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
s3transfer==0.13.0
six==1.17.0
sniffio==1.3.1
snowflake-connector-python==3.16.0
snowflake-sqlalchemy==1.7.6
snowflake._legacy==1.0.1
snowflake.core==1.6.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.41
starlette==0.47.1
tenacity==9.1.2
tiktoken==0.9.0
tomlkit==0.13.3
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
yarl==1.20.1
zstandard==0.23.0
